requirements.txt
Flask
Flask-SQLAlchemy
yfinance
gunicorn
oracledb
python-dotenv
pandas
lxml